requirement.txt
pandas
psycopg2-binary
matplotlib
seaborn
sqlalchemy